refactor(settings): extract provider setup error notice (#2363)
